requirements.txt
python-dotenv==1.0.1
aiogram==3.5.0
together==1.2.1
google-auth==2.29.0
google-api-python-client==2.125.0
colorama==0.4.6
pydub==0.25.1
speechrecognition==3.10.0
ffmpeg-python==0.2.0